fix: do not add id annotation to regcred
We were doing this to try to play around with kustomize's internals. The
id was not necessary, and in fact caused problems because it is a
number. The proper format is for the annotation to be a string. This is
breaking newer version of kubectl (e.g. 1.17).

Remove the id since it is not necessary.