fix: automatically retry on upstream failures
We want to make use of istio's ability to retry http requests on
connection failures. We thought that that was working well, but it turns
out that we were missing a few cases:

1. The default connection timeout is 10 seconds. This is too long, and
   end up hitting it fairly frequently when there's a node reboot.
2. Connection failures after the request has been sent/queued lead to
   retries not happening. This is because the default policy does not
   include those failures (503).

To resolve this, we do two things:

1. Add a destination rule to the ingress/api generators which sets a 2
   second connection timeout. Nothing should take that long, and if it
   does, we'll end up retrying anyway.
2. Explicitly set a retry policy which includes 5xx. This is the default
   policy + 5xx.

Some of our existing config defines their own destination rules. Since
the order of precendence for destination rules isn't the greatest, we
give the ability to disable the destination rule in the IstioIngress
object. To do so, simply add:

```
destination:
  disable: true
```

To the generator spec.