fix: reduce likelihood of outlier ejection
We were having more 503s then we expected reported by our external
healthchecks. After some digging, it looked like this is because envoy
was ejecting one of the two instances of many of the upstreams. When the
node containing the 'up' instance failed, then we lost service. We
really shouldn't have been ejecting at all, let alone for as long as we
were. This change:

 * Decreases the ejection interval so we don't eject for hours at a time
 * Decreases the likelihood of an ejection by increasing the consecutive
   number of failures required to 15 from 2.