feat: allow secrets to have different keys in k8s
Sometimes other tools (such as cert manager) depend on secret values
having specific names. In such cases, it often doens't make sense to
store the secret with that value in sops. So it'd be nice to be able to
map the sops key to a different k8s key, like you can do for example
with the kustomize secret generator.

This allows you to say `a=b` to have the secret be called `a` in
kubernetes despite being called `b` in sops.